requirements.txt
appnope==0.1.0
beautifulsoup4==4.4.1
braintree==3.25.0
click==6.6
csvkit==0.9.1
DateTime==4.0.1
decorator==4.0.6
Django==1.8.5
Flask==0.10.1
Flask-Slack==0.1.5
gnureadline==6.3.3
gunicorn==19.4.5
httplib2==0.9.2
igraph==0.1.11
image==1.4.1
ipython==4.0.1
ipython-genutils==0.1.0
itsdangerous==0.24
jdcal==1.2
Jinja2==2.8
lxml==3.6.0
MarkupSafe==0.23
mlbgame==2.0.2
mock==2.0.0
nflgame==1.2.18
numpy==1.10.2
openpyxl==2.2.0b1
ordereddict==1.1
pandas==0.17.1
path.py==8.1.1
pbr==1.9.1
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
psycopg2==2.6.1
ptyprocess==0.5
pyglet==1.2.4
pyperclip==1.5.13
python-dateutil==2.2
python-dotenv==0.4.0
pytube==6.1.5
pytz==2015.7
qcsv==0.0.6
request==0.0.2
requests==2.10.0
scapy-python3==0.18
simplegeneric==0.8.1
six==1.10.0
slackclient==1.0.0
slacker==0.9.15
SQLAlchemy==1.0.10
traitlets==4.0.0
trueskill==0.4.4
venmo==0.2.2
virtualenv==13.1.2
websocket-client==0.37.0
Werkzeug==0.11.9
xlrd==0.9.4
zope.interface==4.1.3
